refactor: refactor wish list page
